Fix the uninitialized values errors we were getting.
Originally tried declaring the methods in the header file and re-ordering things in mergesort.c, which didnt' work.

It turns out it was a really simply copying error where we had both while loops operating on the second index.
No idea how any tests were passing since that was happening, but everything is happy now.

I've also left the re-ordering changes in because I think it makes the function more java-like and easier to read.