Fix most of the tests for array_merge.
We've seemingly fixed some of the array indices issues, and have actually captured a zero if any of the values in
our values array contained any. (before our comparison to check if we were adding unique values always failed with
zero because malloc and calloc seem to initialize arrays with zeros)

Only the final test: Handle_different_sizes_reversed is failing, where the 0th element (k or length of array)
in our result is 4 where it should be 10. No idea what's going on with that.