Fix the lingering memory leaks in the array_merge_test file.
We were leaking the memory that the tests allocated when they were constructing the values array that was passed to our
functions.

This was fixed by throwing in a for loop that freed the pointers to each internal array of "a" which is an array of
dynamically allocated arrays.

Should be free of all leaks now!